refactor: ResponseFactory 도입 및 응답 생성 로직 통합
- 응답 생성 로직을 담당하는 ResponseFactory 클래스 추가
- cgi::Responder 및 utils/response.hpp의 기능을 ResponseFactory로 통합
- EventHandler 및 각 Builder가 ResponseFactory를 사용하도록 수정
- 불필요해진 Responder 및 response.hpp 파일 삭제
